// src/remote/api/despacho_caja/types/IFriDCjControlVMd3.ts
export interface IFriDCjControlVMd3 {
  idPesaje: number;
  fechaPes: string;
  document?: string;
  placaVeh: string;
  conductorV: string;
  tipoVeh: number;
  tipoTranspPoFs: string;
  transportador?: string;
  persProgramo?: string;
  pesoEntr: number;
  fechaHoraE?: string;
  pesoSali: number;
  fechaHoraS?: string;
  realizo?: string;
  observacion?: string;
  estadoV: string;
  anulado: boolean;
  motiAnul?: string;
  procesado: boolean;
  usuarioV?: string;
  fechaG?: string;
  estacion?: string;
}