// src/remote/api/despacho_caja/types/ITblDCjControlV.ts
export interface ITblDCjControlV {
  idPesaje: number;
  fechaPes: string;
  document?: string;
  placaVeh: string;
  conductorV: string;
  clienteUs: string;
  ciudadDes: string;
  tipoProdu: number;
  tipoVeh: number;
  tipoTranspPoFs: string;
  transportador?: string;
  persProgramo?: string;
  pesoEntr: number;
  fechaHoraE?: string;
  pesoSali: number;
  fechaHoraS?: string;
  realizo?: string;
  observacion?: string;
  estadoV: string;
  anulado: boolean;
  motiAnul?: string;
  procesado: boolean;
  usuarioV?: string;
  fechaG?: string;
  estacion?: string;
}